rework AAA session termination on PDP context closure
Explicitly stop the AAA session whenever a PDP context is terminated.

Leaving the accounting stop handling to the AAA session creates a race
between collecting the accounting data from the data path context and
the removal of that context by the PDP delete function.

This is avoided by collecting the accounting data before the delete.