Initial version of redirector mode for GTP sockets
Now it is possible to create GTP socket in a redirection mode.
The redirection mode means that GTP request will be transferred to
a GTP backed with saving source IP and port. For the GTP backend such
requests look like they came from particular GTP client directly.
Note that it is an initial implementation and the following features
are not implemented:

* IPv6.
* Keep-Alive mechanism with echo requests.
* Memorization of GTP backend in case of retransmission.
* Counter metrics for redirected requests.